refactor(prompt): ♻️ reorganize section sources into modular source directory
Move individual SectionSource implementations from flat module files into a dedicated `sources/` module directory. This improves code organization and maintainability by grouping related source implementations together and reducing the number of top-level modules.

The refactoring includes:
- Creating `sources/mod.rs` to re-export all source implementations
- Moving ActiveGoalSource, ActivePlanSource, CompactionContractSource, and other sources into separate files within the new directory
- Removing `template_sources.rs` which contained implementations now distributed to individual source files
- Updating `registry.rs` and `mod.rs` imports to use the new module structure
- Adding corresponding test file `source_tests.rs` for idempotency and determinism verification
