+++ src-tauri/src/core/prompt/sources/compaction_contract.rs
@@ -1,14 +1,18 @@
 use async_trait::async_trait;
 
-use super::build_context::BuildCx;
-use super::section_source::{FatalError, SectionBody, SectionMeta, SectionOutcome, SectionSource};
-use super::surface::CompactionKind;
-use super::templates::{load_template, parse_front_matter, render_template_strict, TemplateVars};
+use super::super::build_context::BuildCx;
+use super::super::section_source::{
+    FatalError, SectionBody, SectionMeta, SectionOutcome, SectionSource,
+};
+use super::super::surface::CompactionKind;
+use super::super::templates::{
+    load_template, parse_front_matter, render_template_strict, TemplateVars,
+};
 
 const COMPACT_TEMPLATE_REL_PATH: &str = "compaction/compact.md";
-const COMPACT_TEMPLATE_EMBEDDED: &str = include_str!("templates/compaction/compact.md");
+const COMPACT_TEMPLATE_EMBEDDED: &str = include_str!("../templates/compaction/compact.md");
 const MERGE_TEMPLATE_REL_PATH: &str = "compaction/merge.md";
-const MERGE_TEMPLATE_EMBEDDED: &str = include_str!("templates/compaction/merge.md");
+const MERGE_TEMPLATE_EMBEDDED: &str = include_str!("../templates/compaction/merge.md");
 const DECLARED_KEYS: &[&'static str] = &["response_language_line"];
 
 /// Template-backed SectionSource for the CompactionContract section.
@@ -40,7 +44,7 @@ impl SectionSource for CompactionContractSource {
         let raw = load_template(rel_path, embedded);
         let (tmpl, body) = parse_front_matter(&raw).map_err(|e| {
             FatalError::new(
-                super::error_codes::codes::TEMPLATE_NOT_FOUND,
+                super::super::error_codes::codes::TEMPLATE_NOT_FOUND,
                 format!("{}: {}", rel_path, e),
             )
         })?;
@@ -60,7 +64,7 @@ impl SectionSource for CompactionContractSource {
         let vars = TemplateVars::new().insert("response_language_line", response_language_line);
         let rendered = render_template_strict(&body, DECLARED_KEYS, &vars).map_err(|e| {
             FatalError::new(
-                super::error_codes::codes::TEMPLATE_MISSING_KEY,
+                super::super::error_codes::codes::TEMPLATE_MISSING_KEY,
                 format!("{}: {}", rel_path, e),
             )
         })?;
