+++ src-tauri/src/core/prompt/sources/custom_subagent_body.rs
@@ -2,8 +2,10 @@ use async_trait::async_trait;
 
 use crate::core::subagent::SubagentProfile;
 
-use super::build_context::BuildCx;
-use super::section_source::{FatalError, SectionBody, SectionMeta, SectionOutcome, SectionSource};
+use super::super::build_context::BuildCx;
+use super::super::section_source::{
+    FatalError, SectionBody, SectionMeta, SectionOutcome, SectionSource,
+};
 
 // ---------------------------------------------------------------------------
 // SubagentBodySource: loads template-backed subagent body for Explore/Review;
@@ -21,17 +23,17 @@ impl SectionSource for SubagentBodySource {
     async fn build(&self, cx: &BuildCx<'_>) -> Result<SectionOutcome, FatalError> {
         match cx.helper_profile {
             Some(SubagentProfile::Explore) => {
-                let template = include_str!("templates/subagent/explore.md");
+                let template = include_str!("../templates/subagent/explore.md");
                 let (_tmpl, body) =
-                    super::templates::parse_front_matter(template).map_err(|e| {
+                    super::super::templates::parse_front_matter(template).map_err(|e| {
                         FatalError::new("template.parse", format!("subagent/explore.md: {e}"))
                     })?;
                 // No template vars needed for static persona prompts
-                let vars = super::templates::TemplateVars::new();
-                let rendered = super::templates::render_template_strict(&body, &[], &vars)
+                let vars = super::super::templates::TemplateVars::new();
+                let rendered = super::super::templates::render_template_strict(&body, &[], &vars)
                     .map_err(|e| {
-                        FatalError::new("template.render", format!("subagent/explore.md: {e}"))
-                    })?;
+                    FatalError::new("template.render", format!("subagent/explore.md: {e}"))
+                })?;
                 Ok(SectionOutcome::Produced(SectionBody {
                     markdown: rendered,
                     meta: SectionMeta {
@@ -41,16 +43,16 @@ impl SectionSource for SubagentBodySource {
                 }))
             }
             Some(SubagentProfile::Review) => {
-                let template = include_str!("templates/subagent/review.md");
+                let template = include_str!("../templates/subagent/review.md");
                 let (_tmpl, body) =
-                    super::templates::parse_front_matter(template).map_err(|e| {
+                    super::super::templates::parse_front_matter(template).map_err(|e| {
                         FatalError::new("template.parse", format!("subagent/review.md: {e}"))
                     })?;
-                let vars = super::templates::TemplateVars::new();
-                let rendered = super::templates::render_template_strict(&body, &[], &vars)
+                let vars = super::super::templates::TemplateVars::new();
+                let rendered = super::super::templates::render_template_strict(&body, &[], &vars)
                     .map_err(|e| {
-                        FatalError::new("template.render", format!("subagent/review.md: {e}"))
-                    })?;
+                    FatalError::new("template.render", format!("subagent/review.md: {e}"))
+                })?;
                 Ok(SectionOutcome::Produced(SectionBody {
                     markdown: rendered,
                     meta: SectionMeta {
