+++ src-tauri/src/core/prompt/sources/skills.rs
@@ -2,12 +2,16 @@ use async_trait::async_trait;
 
 use crate::extensions::{ConfigScope, ExtensionsManager};
 
-use super::build_context::BuildCx;
-use super::section_source::{FatalError, SectionBody, SectionMeta, SectionOutcome, SectionSource};
-use super::templates::{load_template, parse_front_matter, render_template_strict, TemplateVars};
+use super::super::build_context::BuildCx;
+use super::super::section_source::{
+    FatalError, SectionBody, SectionMeta, SectionOutcome, SectionSource,
+};
+use super::super::templates::{
+    load_template, parse_front_matter, render_template_strict, TemplateVars,
+};
 
 const TEMPLATE_REL_PATH: &str = "skills_usage.md";
-const TEMPLATE_EMBEDDED: &str = include_str!("templates/skills_usage.md");
+const TEMPLATE_EMBEDDED: &str = include_str!("../templates/skills_usage.md");
 const DECLARED_KEYS: &[&'static str] = &["skills_list"];
 
 /// Template-backed SectionSource for the Skills section.
@@ -29,7 +33,7 @@ impl SectionSource for SkillsSource {
             Ok(skills) => skills,
             Err(e) => {
                 return Ok(SectionOutcome::SoftFailed {
-                    code: super::error_codes::codes::SKILLS_LOAD_FAILED,
+                    code: super::super::error_codes::codes::SKILLS_LOAD_FAILED,
                     error: Box::new(std::io::Error::new(
                         std::io::ErrorKind::Other,
                         e.to_string(),
@@ -67,7 +71,7 @@ impl SectionSource for SkillsSource {
         let raw = load_template(TEMPLATE_REL_PATH, TEMPLATE_EMBEDDED);
         let (_tmpl, body) = parse_front_matter(&raw).map_err(|e| {
             FatalError::new(
-                super::error_codes::codes::TEMPLATE_NOT_FOUND,
+                super::super::error_codes::codes::TEMPLATE_NOT_FOUND,
                 format!("{}: {}", TEMPLATE_REL_PATH, e),
             )
         })?;
@@ -75,7 +79,7 @@ impl SectionSource for SkillsSource {
         let vars = TemplateVars::new().insert_user_text("skills_list", skills_list);
         let rendered = render_template_strict(&body, DECLARED_KEYS, &vars).map_err(|e| {
             FatalError::new(
-                super::error_codes::codes::TEMPLATE_MISSING_KEY,
+                super::super::error_codes::codes::TEMPLATE_MISSING_KEY,
                 format!("{}: {}", TEMPLATE_REL_PATH, e),
             )
         })?;
